requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic[email]
python-dotenv